gl-renderer: Build as a loadable module
In some embedded use cases, the time spent loading (E)GL libs into
memory is a considerable hit. This patch changes the gl renderer
interface so that it can be compiled as a separate module, that is
dlopen'd by the backends. This is a first step towards having the
GL renderer initialization delayed, while still displaying some sort
of splash screen during that time.

The interface is changed so that there is only one symbom that needs
to be dlsym'd. This is a structure with function pointers to all the
entry points. This has a side effect of simplifying the header file
for the case EGL is not present.

DRM, X11 and Wayland backends are converted to the new interface.