compositor-drm: Add a splash screen interface
Add a (hackish) splash screen interface. The surface set as the splash
with that interface will be shown until the splash interface receives
the done request, at which point it will launch desktop-shell.

The splash-client is just a modified version of simple-shm, modified to
use the splash interface instead of wl_shell. After running for a few
seconds it will send the done request.

Change-Id: I5ae6e511d3f68787ead2104e630a592487162024