compositor-drm: Improve video initialization times
If we are using the DRM backend, and a fixed video mode was
specified in "weston.ini", try to discover DRM modes until
a matching one is found ; then stop.

Reverse the modes loop order ; from the highest resolution
to the lowest now, which is a more frequent use case.

Change-Id: I9832e4081df0d8e94edb997359f447536abe4280
Signed-off-by: Manuel Bachmann <manuel.bachmann@open.eurogiciel.org>