toytoolkit: always mask the frame, even in maximized mode
When a toytoolkit window was maximized, we avoided some
processing by not setting the transparency and mask of the
underlying frame, because the visible result were the
shadows -which are not supposed to be shown in this state.

However, with some GL implementations (ex: Tizen Emulator
with YaGL), this defines a totally transparent window which
is then not shown, leaving the impression that the client
is hung.

Bug-Tizen: TC-2276

Change-Id: I9459e5f177461746efbf3c0a7e1610ac441d8388
Signed-off-by: Manuel Bachmann <manuel.bachmann@open.eurogiciel.org>